refactor(rankings): optimize global rank retrieval for individual users
- Replaced full leaderboard materialization with a targeted query for a single user's rank.
- Changed `RankingRepository`'s `score` field from `Long` to `Double` for precision.
- Updated `RankingEntryResponse` DTO to reflect the `score` type change.
- Improved inline documentation for clarity and maintainability.

+++ backend/src/main/kotlin/org/cityxplore/backend/social/rankings/dto/RankingEntryResponse.kt
@@ -18,7 +18,7 @@ data class RankingEntryResponse(
     val userId: UUID,
     val username: String,
     val avatarUrl: String?,
-    val score: Long,
+    val score: Double,
     val totalPoisDiscovered: Int,
     val totalDistance: Double,
     val totalAchievementPoints: Int,
